requirements.txt
mujoco-py==0.5.7
gym==0.7.4
tensorflow-gpu==1.12.0